fix migrations for existing tables

+++ database/migrations/2025_06_09_203220_privatais_katalogs.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('privatais_katalogs')) {
+            return;
+        }
         Schema::create('privatais_katalogs', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
